+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog records changes to stable releases since 1.50.2. "TBA" changes he
 
 ## TBA
 
+- feat: implement 'start debugging and stop on entry' command/keybinding ([ref](https://github.com/microsoft/vscode/issues/49855))
 - fix: pretty print not working when evaling sources ([#929](https://github.com/microsoft/vscode-js-debug/issues/929))
 - fix: browser debugging in remote not working on some Linux systems ([#908](https://github.com/microsoft/vscode-js-debug/issues/908))
 - fix: edge not launching if VS Code is run in admin mode on windows ([ref](https://github.com/microsoft/vscode/issues/117005))